+++ packages/xyd-cli/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @xyd-js/cli
 
+## 0.1.0-xyd.44
+
+### Patch Changes
+
+- cli
+
+## 0.1.0-xyd.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/documan@0.1.0-xyd.34
+
+## 0.1.0-xyd.42
+
+### Patch Changes
+
+- cli
+
+## 0.1.0-xyd.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/documan@0.1.0-xyd.33
+  - @xyd-js/host@0.1.0-xyd.3
+
+## 0.1.0-xyd.40
+
+### Patch Changes
+
+- update cli
+
+## 0.1.0-xyd.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/documan@0.1.0-xyd.32
+
+## 0.1.0-xyd.38
+
+### Patch Changes
+
+- update cli
+
 ## 0.1.0-xyd.37
 
 ### Patch Changes